GPU accelerate arrow transition, restyle team hover
Hmm hardware acceleration should be automatic with any transform, but it
does seem to have improved my Chrome lag issue. Still not perfect, but
it might just be my shitty Intel graphics.

Edge and FF were never affected to begin with.

And matched the team section to the site color scheme.